Move map into a file
